requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
pymysql